// src/app/interfaces/page.interface.ts
export interface Page {
  id: number;
  linkButtonAndroid: string;
  linkButtonIPhone: string;
  linkButtonWinPhone: string;
  pageBackgroundColor: string;
  pageName: string;
  userId: number;
  userHtml: string;
  heading: string;
  headingColor: string;
  headingFontId: number;
  androidCompat: boolean;
  iphoneCompat: boolean;
  winPhoneCompat: boolean;
  description: string;
  isAdvancedPage: boolean;
  smallPictureId: number;
  bannerImage1Id: number;
}